Refactored to use a Task and Command approach, seems to be more stable and cleaner
