+++ include/global_state.h
@@ -9,46 +9,39 @@
 #include <Arduino.h>
 #include <vector>
 
-struct NetworkState {
-    bool wifi_connected = false;
-    String ip = "0.0.0.0";
-    String selected_ssid;
-    String selected_pass;
-    std::vector<String> found_ssids;
-};
+// --- Command Bitmasks ---
+#define CMD_WIFI_SCAN    (1 << 0)
+#define CMD_WIFI_CONN    (1 << 1)
+#define CMD_WIFI_RESET   (1 << 2)
 
-struct TransitionFlags {
-    // Scanning Flags
-    volatile bool start_scan_trigger = false;
-    volatile bool scan_complete = false;
-
-    // Connection Flags
-    volatile bool is_connecting = false;
-    volatile bool start_connect_trigger = false;
-    volatile bool failed_to_connect_trigger = false;
-    volatile bool show_success_trigger = false;
-
-    // Error Flags
-    volatile bool fatal_error_trigger = false;
-
-    uint32_t success_shown_at = 0;
-};
+// --- Task Handles ---
+extern TaskHandle_t systemTaskHandle;
 
 struct SystemState {
     bool setup_complete = false;
     bool spotify_linked = false;
-    String spotify_auth_url = "";
-
-    // Player Data
-    String current_track = "Idle";
-    String artist = "Waiting...";
-    int volume = 50;
-    bool is_playing = false;
 };
 
-// Declare a global instance that all files can see
-extern SystemState deviceState;
+
+enum WifiStatus {
+    WIFI_IDLE,
+    WIFI_SCANNING,
+    WIFI_SCAN_RESULTS,
+    WIFI_CONNECTING,
+    WIFI_CONNECTED,
+    WIFI_ERROR
+};
+struct NetworkState {
+    WifiStatus status = WIFI_IDLE;
+    bool wifi_connected = false;
+    String ip = "0.0.0.0";
+
+    String selected_ssid = "";
+    String selected_pass = "";
+    std::vector<String> found_ssids;
+};
+
+extern SystemState systemState;
 extern NetworkState networkState;
-extern TransitionFlags transitionFlags;
 
 #endif //GLOBAL_STATE_H
