+++ platformio.ini
@@ -2,6 +2,8 @@
 platform = espressif32@6.6.0
 board = esp32-s3-devkitc-1
 framework = arduino
+monitor_speed = 115200
+upload_speed = 115200
 
 board_build.arduino.memory_type = qio_opi
 board_build.flash_mode = qio
