+++ src/spotify/SpotifyManager.cpp
@@ -1,0 +1,5 @@
+//
+// Created by Harry Skerritt on 06/01/2026.
+//
+
+#include "SpotifyManager.h"
