+++ src/spotify/SpotifyManager.h
@@ -1,0 +1,16 @@
+//
+// Created by Harry Skerritt on 06/01/2026.
+//
+
+#ifndef SPOTIFYMANAGER_H
+#define SPOTIFYMANAGER_H
+
+
+
+class SpotifyManager {
+
+};
+
+
+
+#endif //SPOTIFYMANAGER_H
